add safe param convert

+++ twod_gs/Method/point_utils.py
@@ -2,20 +2,20 @@ import torch
 
 
 def depths_to_points(view, depthmap):
-    c2w = (view.world_view_transform.T).inverse()
+    c2w = (view.world_view_transform.T).float().inverse()
     W, H = view.image_width, view.image_height
     ndc2pix = torch.tensor([
         [W / 2, 0, 0, (W) / 2],
         [0, H / 2, 0, (H) / 2],
         [0, 0, 0, 1]]).float().cuda().T
-    projection_matrix = c2w.T @ view.full_proj_transform
+    projection_matrix = c2w.T @ view.full_proj_transform.float()
     intrins = (projection_matrix @ ndc2pix)[:3,:3].T
     
     grid_x, grid_y = torch.meshgrid(torch.arange(W, device='cuda').float(), torch.arange(H, device='cuda').float(), indexing='xy')
     points = torch.stack([grid_x, grid_y, torch.ones_like(grid_x)], dim=-1).reshape(-1, 3)
     rays_d = points @ intrins.inverse().T @ c2w[:3,:3].T
     rays_o = c2w[:3,3]
-    points = depthmap.reshape(-1, 1) * rays_d + rays_o
+    points = depthmap.float().reshape(-1, 1) * rays_d + rays_o
     return points
 
 def depth_to_normal(view, depth):
